add some start battle feature

+++ Assets/Scripts/Moves/MoveSO.cs
@@ -9,7 +9,7 @@ public class MoveSO : ScriptableObject
     [SerializeField] private PokemonType moveType;
     [SerializeField] private int movePower;
     [SerializeField] private int moveAccuracy;
-    [SerializeField] private float movePp;
+    [SerializeField] private int movePp;
     [SerializeField] private bool isMakeContact;
     [SerializeField] private int movePriority;
     [SerializeField] private MoveCategory moveCategory;
@@ -22,7 +22,7 @@ public class MoveSO : ScriptableObject
     public string MoveName { get => moveName; set => moveName = value; }
     public PokemonType MoveType { get => moveType; set => moveType = value; }
     public int MovePower { get => movePower; set => movePower = value; }
-    public float MovePp { get => movePp; set => movePp = value; }
+    public int MovePp { get => movePp; set => movePp = value; }
     public bool IsMakeContact { get => isMakeContact; set => isMakeContact = value; }
     public int MovePriority { get => movePriority; set => movePriority = value; }
     public int MoveAccuracy { get => moveAccuracy; set => moveAccuracy = value; }
